+++ neat/src/networks/connections.rs
@@ -11,13 +11,13 @@ pub struct Connection<'a> {
     id: Innovation,
     input: Rc<RefCell<Node<'a>>>,
     output: Rc<RefCell<Node<'a>>>,
-    weight: f64,
+    weight: f32,
 }
 
 impl<'a> Connection<'a> {
     /// Create a new connection between the specified
     /// input and output nodes.
-    pub fn new(input_id: Rc<RefCell<Node<'a>>>, output_id: Rc<RefCell<Node<'a>>>, weight: f64) {
+    pub fn new(input_id: Rc<RefCell<Node<'a>>>, output_id: Rc<RefCell<Node<'a>>>, weight: f32) {
         todo!()
     }
